% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ApuliaAQ}
\alias{ApuliaAQ}
\title{Apulia Air Quality Data}
\format{
\subsection{\code{ApuliaAQ}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{AirQualityStation}{Station code}
\item{time}{Date of measurement}
\item{Longitude, Latitude}{Geographic coordinates using the WGS-84 reference system}
\item{Altitude}{Altitude of the measurement station (m)}
\item{AirQualityStationType}{predominant type of emission sources in the station's vicinity}
\item{AirQualityStationArea}{type of area surrounding the station}
\item{AQ_mean_NO2, AQ_mean_PM10, AQ_mean_PM2.5}{Air pollutant concentrations (micrograms per cubic meter)}
\item{CL_blh}{Daily mean of the height of the atmosphere boundary layer (m)}
\item{CL_lai_hv}{Daily fixed value of high vegetation leaf area index (m2/m2)}
\item{CL_lai_lv}{Daily fixed value of low vegetation leaf area index (m2/m2)}
\item{CL_rh}{Daily mean of relative humidity (percent)}
\item{CL_ssr}{Daily maximum of surface solar radiation (J/m2)}
\item{CL_t2m}{Daily mean of air temperature at 2 meters (Celsius)}
\item{CL_tp}{Daily cumulative total precipitation (m)}
\item{CL_winddir}{Daily mode of wind direction (1=N, 2=E, 3=S, 4=W)}
\item{CL_windspeed}{Daily mean of wind speed (m/s)}
}
}
}
\source{
\url{https://doi.org/10.5281/zenodo.15699805}
}
\usage{
ApuliaAQ
}
\description{
Daily measurements of some air pollutant concentrations,
recorded by ground-based monitoring stations in Apulia (Italy) in 2022,
along with related climate variables. This is a subset of data from the GRINS AQCLIM dataset (see link below).
}
\keyword{datasets}
