% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{coef.stdglm}
\alias{coef.stdglm}
\title{Coefficients in \code{stdglm} objects}
\usage{
\method{coef}{stdglm}(object, type = "fitted", ...)
}
\arguments{
\item{object}{An \code{stdglm} object.}

\item{type}{A character string indicating the type of coefficients. Options are:
\itemize{
\item \code{overall}: Overall effects of varying coefficients.
\item \code{tvc}: Temporal effects of varying coefficients.
\item \code{svc}: Spatial effects of varying coefficients.
\item \code{stvc}: Spatio-temporal effects of varying coefficients.
\item \code{gamma}: Effects of covariates specified in the input \code{Z} (see \code{\link{stdglm}}).
}}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A dataframe with posterior mean and 95\% credible intervals.
}
\description{
Extracts coefficients from a \code{stdglm} object.
}
\details{
Extracts posterior mean and 95\% credible intervals of the coefficients according to their type.
}
\examples{
\dontrun{
# Assuming `mod` is a fitted stdglm object
print(coef(mod, 'overall'))
print(coef(mod, 'gamma'))
betaST_post = coef(mod, 'stvc') # returns a data.frame
head(betaST_post)
}
}
\seealso{
\code{\link{stdglm}}
}
