% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{fitted.stdglm}
\alias{fitted.stdglm}
\title{Fitted values for \code{stdglm} objects}
\usage{
\method{fitted}{stdglm}(object, ...)
}
\arguments{
\item{object}{An \code{stdglm} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A matrix of fitted values.
}
\description{
Extracts the fitted values from an \code{stdglm} object.
}
\details{
Extracts the posterior mean of the predictive distribution for the observed data points.
}
\examples{
\dontrun{
# Assuming `mod` is a fitted stdglm object
fitted_values <- fitted(mod)
summary(t(fitted_values[1:5,])) # Extract temporal summaries for the first 5 locations
}
}
\seealso{
\code{\link{stdglm}}
}
