% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{plot.stdglm}
\alias{plot.stdglm}
\title{Plotting functions for \code{stdglm} objects}
\usage{
\method{plot}{stdglm}(x, type = "fitted", ...)
}
\arguments{
\item{x}{An \code{stdglm} object.}

\item{type}{A character string indicating the type of plot to generate. Options are:
\itemize{
\item \code{tvc}: Temporal effects of varying coefficients.
\item \code{svc}: Spatial effects of varying coefficients.
\item \code{stvc}: Spatio-temporal effects of varying coefficients.
\item \code{fitted}: Fitted values of the model.
}}

\item{...}{Additional arguments passed to the plotting functions.}
}
\value{
A ggplot object or a list of ggplot objects, depending on the type of plot.
}
\description{
Plotting functions for \code{stdglm} objects.
}
\details{
The \code{plot.stdglm} function dispatches to the appropriate plotting function based on the \code{type} argument.
}
\examples{
\dontrun{
# Assuming `mod` is a fitted stdglm object
plot(mod, type = 'tvc')  # Plot temporal effects
plot(mod, type = 'svc', Coo_sf_obs, region)  # Plot spatial effects, 
# where `Coo_sf_obs` is a spatial object with observed coordinates and 
# `region` is a spatial region object, both from the `sf` package.
plot(mod, type = 'stvc', 1:4)  # Plot spatio-temporal effects
plot(mod, type = 'fitted', Y, 1)  # Plot fitted values
}
}
\seealso{
\code{\link{stdglm}}, \code{\link[ggplot2]{ggplot}}
}
\keyword{plot}
