% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{predict.stdglm}
\alias{predict.stdglm}
\title{Retrieve predictions for \code{stdglm} objects}
\usage{
\method{predict}{stdglm}(object, type = "response_mat", Coo_sf_pred = NULL, ...)
}
\arguments{
\item{object}{An \code{stdglm} object.}

\item{type}{A character string indicating the type of coefficients. Options are:
\itemize{
\item \code{response_mat}: Posterior mean of the predictive distribution for the response variable. Returns a \emph{p_new}-by-\emph{t_new} matrix (default).
\item \code{response_df}: Mean, standard deviation and 95 \% credible interval for the response variable (default). Returns a dataframe.
\item \code{tvc}: Mean, standard deviation and 95 \% credible interval for the temporal effects of varying coefficients.
\item \code{svc}: Mean, standard deviation and 95 \% credible interval for the spatial effects of varying coefficients.
\item \code{stvc}: Mean, standard deviation and 95 \% credible interval for the spatio-temporal effects of varying coefficients.
}}

\item{Coo_sf_pred}{A simple feature object from package \code{sf} with the prediction points, whose geometry is used for spatial effects (optional).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Either a matrix or a dataframe or an \code{sf} object with posterior mean and 95\% credible interval bounds. The function returns \code{NULL} if predictions are not available in the \code{stdglm} object.
}
\description{
Extracts out-of-sample predictions from an \code{stdglm} object.
}
\details{
Returns the posterior mean of the predictive distribution and the associated 95\% credible intervals for the out-of-sample data points, and for the random variable specified in the input \code{type}.
}
\examples{
\dontrun{
# Assuming `mod` is a fitted stdglm object
predictions <- predict(mod)
colMeans(predictions) # Get the average predictions across all locations
pred_df <- predict(mod, type = 'response_df')
head(pred_df)
}
}
\seealso{
\code{\link{stdglm}}, \code{\link[sf]{sf}}
}
