% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{fitted.stdglm}
\alias{fitted.stdglm}
\title{Fitted values for \code{stdglm} objects}
\usage{
\method{fitted}{stdglm}(object, ...)
}
\arguments{
\item{object}{A \code{stdglm} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A matrix of fitted values.
}
\description{
Extracts the fitted values from a \code{stdglm} object.
}
\details{
Extracts the posterior mean of the predictive distribution for the observed data points.
}
