% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdglm.R
\name{stdglm}
\alias{stdglm}
\title{MCMC Algorithm}
\usage{
stdglm(
  y,
  X,
  Z = NULL,
  offset = NULL,
  point.referenced = TRUE,
  random.walk = FALSE,
  W,
  W_pred = NULL,
  W_cross = NULL,
  X_pred = NULL,
  Z_pred = NULL,
  offset_pred = NULL,
  nrep = 100,
  nburn = 100,
  thin = 1,
  print.interval = 100,
  prior = NULL,
  keepY = TRUE,
  keepLogLik = TRUE,
  last_run = NULL
)
}
\arguments{
\item{y}{Either a \emph{p}-by-\emph{t} matrix corresponding to the response variable, where \emph{p} is the number of spatial locations and \emph{t} is the number of time points.}

\item{X}{A \emph{p}-by-\emph{t}-by-\emph{ncovx} array corresponding to the covariates whose effects should vary across space and time.}

\item{Z}{A \emph{p}-by-\emph{t}-by-\emph{ncovz} array corresponding to the covariates whose effects are constant across space and time. Defaults to \code{NULL}.}

\item{offset}{A \emph{p}-by-\emph{t} matrix corresponding to the offset term. If \code{NULL}, it is set to zero.}

\item{point.referenced}{A logical indicating whether the data are point-referenced (TRUE) or areal (FALSE). Default is TRUE. If FALSE, predictions are not performed.}

\item{random.walk}{A logical indicating whether the temporal dynamic should be modeled as a random walk (TRUE) or a first-order autoregressive process (FALSE). Default is FALSE.}

\item{W}{A \emph{p}-by-\emph{p} matrix corresponding to the spatial weights matrix. If \code{point.referenced} is TRUE, the distance matrix among the observed locations should be provided. If \code{point.referenced} is FALSE, the 0/1 adjacency matrix should be provided.}

\item{W_pred}{A \emph{p}-by-\emph{p} matrix corresponding to the spatial weights matrix for the prediction locations. If \code{NULL}, predictions are not performed.}

\item{W_cross}{A \emph{p}-by-\emph{p_new} matrix corresponding to the cross spatial weights matrix between the observed and prediction locations.}

\item{X_pred}{A \emph{p_new}-by-\emph{t_new}-by-\emph{ncovx} array corresponding to the covariates with varying coefficients for predictions, where \emph{p_new} is the number of prediction locations and \emph{t_new=t+h_ahead}, \emph{h_ahead}>=0, is the number of time points for which predictions are to be made.}

\item{Z_pred}{A \emph{p_new}-by-\emph{t_new}-by-\emph{ncovz} array corresponding to the covariates with constant coefficients for predictions. Defaults to \code{NULL}.}

\item{offset_pred}{A \emph{p_new}-by-\emph{t_new} matrix corresponding to the offset term for predictions. If \code{NULL}, it is set to zero.}

\item{nrep}{An integer indicating the number of iterations to keep after burn-in.}

\item{nburn}{An integer indicating the number of iterations to discard as burn-in.}

\item{thin}{An integer indicating the thinning value. Default is 1.}

\item{print.interval}{An integer indicating the interval at which to print progress messages.}

\item{prior}{A named list containing the hyperparameters of the model. If \code{NULL}, default non-informative hyperparameters are used. See details below.}

\item{keepY}{A logical indicating whether to keep the response variable in the output. Default is TRUE.}

\item{keepLogLik}{A logical indicating whether to keep the log-likelihood in the output. Default is TRUE.}

\item{last_run}{An optional list containing the output from a previous run of the function, which can be used to restore the state of the sampler. Default is \code{NULL}.}
}
\value{
A named list containing \code{out} and \code{ave}. \code{ave} stores posterior summaries, while \code{out} stores the MCMC samples.
}
\description{
This function executes the Markov Chain Monte Carlo (MCMC) algorithm for STDGLM.
}
\details{
\code{prior} must be a named list with these elements:
\describe{
\item{V_beta_0}{A scalar defining the prior variance for the initial state of the time-varying coefficients, related to the covariates in \code{X}.}
\item{V_gamma}{A scalar defining the prior variance for the initial state of the constant coefficients, related to the covariates in \code{Z}. Assign a value even if \code{Z} is \code{NULL}.}
\item{a1}{A scalar defining the inverse-gamma prior shape for the temporal variance of the time-varying coefficients.}
\item{b1}{A scalar defining the inverse-gamma prior rate for the temporal variance of the time-varying coefficients.}
\item{s2_a}{A scalar defining the inverse-gamma prior shape for the measurement error variance.}
\item{s2_b}{A scalar defining the inverse-gamma prior rate for the measurement error variance.}
}
}
\references{
insert here.
}
